Link and Projects Section Update
All links now changed and working properly whether
it be internal or external for different pages,
social media, personal links, and even project links.
The reason the project links weren't working was
because the z-index was -1 and the z-index of the body
is already set to 0. Meaning the links themselves were
under the body and not able to be clicked. Fixed this
by using different position values for divs as a
 static div is set to be under an element with
relative position The reason the footer links weren't
working was because they were buttons and not anchors.