Add Text Change On Hover
Project section now has text changing when
hovering over the corresponding project and
will give a description of that project.
Before adding all of this to the other planned
transitions, I want to learn fade-ins so that I
can easily copy this to all the other parts of
the website.